JHA: Added page content mixin.
