--- jha/src/helpers/dom.js
@@ -1,21 +1,0 @@
-export const getPageContentDeltaHeight = () => {
-  const headerEl = document.querySelector('header');
-  const footerEl = document.querySelector('footer');
-  const stepperEl = document.querySelector('.progress-bar-component');
-  const continueBarEl = document.querySelector('.continue-bar');
-  let delta = 0;
-
-  if (headerEl) {
-    delta += headerEl.offsetHeight;
-  }
-  if (footerEl) {
-    delta += footerEl.offsetHeight;
-  }
-  if (stepperEl) {
-    delta += stepperEl.offsetHeight;
-  }
-  if (continueBarEl) {
-    delta += continueBarEl.offsetHeight;
-  }
-  return delta;
-};
