+++ jha/src/views/form-a/FormPage.vue
@@ -38,11 +38,12 @@ import {
   RESET_FORM,
 } from '@/store/modules/form-a-module';
 import logService from '@/services/log-service';
-import { getPageContentDeltaHeight } from '@/helpers/dom';
 import { PageContent } from 'common-lib-vue';
+import pageContentMixin from '@/mixins/page-content-mixin';
 
 export default {
   name: 'MainFormPage',
+  mixins: [pageContentMixin],
   components: {
     ContinueBar,
     PageContent,
@@ -50,7 +51,6 @@ export default {
   data: () => {
     return {
       isPageLoaded: false,
-      pageContentDeltaHeight: 0,
     };
   },
   created() {
@@ -63,10 +63,6 @@ export default {
       formARoutes.FORM_PAGE.path,
       formARoutes.FORM_PAGE.title
     );
-
-    this.$nextTick(() => {
-      this.pageContentDeltaHeight = getPageContentDeltaHeight();
-    });
   },
   validations() {
     const validations = {};
